# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Building Maintenance API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/residents/{id}": {
      "get": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Get a resident by ID",
        "description": "Returns a resident by their unique identifier.",
        "operationId": "getOne_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resident found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Update a resident",
        "description": "Updates an existing resident by their ID.",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Resident update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentUpdateCommandDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Resident updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Delete a resident",
        "description": "Deletes a resident by their ID.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resident deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/residents": {
      "get": {
        "tags": [
          "Resident Management"
        ],
        "summary": "List all residents",
        "description": "Returns a list of all residents.",
        "operationId": "listAll_1",
        "responses": {
          "200": {
            "description": "List of residents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResidentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Create a new resident",
        "description": "Creates a new resident with the provided details.",
        "operationId": "create_1",
        "requestBody": {
          "description": "Resident creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentCreateCommandDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Resident created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResidentUpdateCommandDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string"
          },
          "unitNumber": {
            "type": "string",
            "minLength": 1
          },
          "unitBlock": {
            "type": "string",
            "minLength": 1
          },
          "residentialComplexId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "email",
          "fullName",
          "unitBlock",
          "unitNumber"
        ]
      },
      "ResidentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "unitNumber": {
            "type": "string"
          },
          "unitBlock": {
            "type": "string"
          },
          "residentialComplex": {
            "$ref": "#/components/schemas/ResidentialComplexDto"
          }
        }
      },
      "ResidentialComplexDto": {
        "type": "object",
        "description": "Summary information about a residential complex.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the residential complex."
          },
          "name": {
            "type": "string",
            "description": "Name of the residential complex."
          },
          "address": {
            "type": "string",
            "description": "Address of the residential complex."
          },
          "city": {
            "type": "string",
            "description": "City where the residential complex is located."
          },
          "postalCode": {
            "type": "string",
            "description": "Postal code of the residential complex."
          }
        }
      },
...
      "ResidentCreateCommandDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "unitNumber": {
            "type": "string",
            "minLength": 1
          },
          "unitBlock": {
            "type": "string",
            "minLength": 1
          },
          "residentialComplexId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "email",
          "fullName",
          "password",
          "residentialComplexId",
          "unitBlock",
          "unitNumber"
        ]
      },
...
      "DeleteResponseDto": {
        "type": "object",
        "description": "DTO returned after deleting a resource, containing a message about the operation result.",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message describing the result of the delete operation.",
            "example": "Residential complex deleted successfully."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}